requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
filelock==3.8.0
ftfy==6.1.1
huggingface-hub==0.10.0
idna==3.4
joblib==1.2.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
